Create project for bridge test
