+++ Lombiq.Hosting.MediaTheme.sln
@@ -10,6 +10,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lombiq.Hosting.MediaTheme.T
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lombiq.Hosting.MediaTheme.Targets", "Lombiq.Hosting.MediaTheme.Targets\Lombiq.Hosting.MediaTheme.Targets.csproj", "{49FF1549-F170-4B17-90C7-3F242A4BEBD6}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lombiq.Hosting.MediaTheme.Bridge.Tests.UI", "Lombiq.Hosting.MediaTheme.Bridge.Tests.UI\Lombiq.Hosting.MediaTheme.Bridge.Tests.UI.csproj", "{1B12DA93-22B4-4698-8FD2-D1FA77326A1E}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -36,5 +38,9 @@ Global
 		{49FF1549-F170-4B17-90C7-3F242A4BEBD6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{49FF1549-F170-4B17-90C7-3F242A4BEBD6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{49FF1549-F170-4B17-90C7-3F242A4BEBD6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1B12DA93-22B4-4698-8FD2-D1FA77326A1E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1B12DA93-22B4-4698-8FD2-D1FA77326A1E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1B12DA93-22B4-4698-8FD2-D1FA77326A1E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1B12DA93-22B4-4698-8FD2-D1FA77326A1E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
